+++ basicsr/archs/edsr_arch.py
@@ -1,8 +1,7 @@
 import torch
 from torch import nn as nn
 
-from basicsr.models.archs.arch_util import (ResidualBlockNoBN, Upsample,
-                                            make_layer)
+from basicsr.archs.arch_util import ResidualBlockNoBN, Upsample, make_layer
 from basicsr.utils.registry import ARCH_REGISTRY
