+++ basicsr/models/losses/losses.py
@@ -4,7 +4,7 @@ from torch import autograd as autograd
 from torch import nn as nn
 from torch.nn import functional as F
 
-from basicsr.models.archs.vgg_arch import VGGFeatureExtractor
+from basicsr.archs.vgg_arch import VGGFeatureExtractor
 from basicsr.models.losses.loss_util import weighted_loss
 from basicsr.utils.registry import LOSS_REGISTRY
