+++ basicsr/models/sr_model.py
@@ -3,10 +3,10 @@ from collections import OrderedDict
 from os import path as osp
 from tqdm import tqdm
 
+from basicsr.archs import build_network
 from basicsr.metrics import calculate_metric
 from basicsr.utils import get_root_logger, imwrite, tensor2img
 from basicsr.utils.registry import MODEL_REGISTRY
-from .archs import build_network
 from .base_model import BaseModel
 from .losses import build_loss
